+++ Main/src/main/java/com/rubbertranslator/mvp/modules/log/LoggerManager.java
@@ -1,4 +1,7 @@
 package com.rubbertranslator.mvp.modules.log;
+import com.rubbertranslator.utils.OSTypeUtil;
+import sun.awt.OSInfo;
+
 import java.io.File;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
@@ -13,7 +16,16 @@ import java.util.logging.SimpleFormatter;
  * date 2020/5/8 10:49
  */
 public class LoggerManager {
-    private static final String logFilePath = System.getProperty("user.dir") + "/RubberTranslator/log";
+    private static final String logFilePath ;
+
+    static{
+        if(OSTypeUtil.isMac()){
+            logFilePath = System.getProperty("user.home") + "/RubberTranslator/log";
+        }else{
+            logFilePath = System.getProperty("user.dir") + "/RubberTranslator/log";
+        }
+    }
+
 
     private static Logger fileLogger = Logger.getLogger("com.rubbertranslator");
 
@@ -36,7 +48,7 @@ public class LoggerManager {
             dir.getParentFile().mkdirs();
         }else{
             File[] logs = dir.getParentFile().listFiles();
-            if(logs != null){
+            if(logs != null && logs.length > 10){
                 for(File log: logs){
                     if(log.exists()) log.delete();
                 }
